chore: 폴더 구조 수정 및 클래스 네임 수정
ZergBug -> ZergUnit

+++ lib/231206/starcraft/interface/scv_repairable.dart
@@ -1,4 +1,4 @@
 abstract interface class SCVRepairable {
-  double get hp;
-  set hp(double value);
+  int get hp;
+  set hp(int value);
 }
